+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [1.3.2] - 2021-08-13
+
 ### Fixed
 - MD Schema
 ## [1.3.1] - 2021-08-13